UPDATED TYPES
Updated type effectiveness to gen 9
Changed Bird type to Type less, because of the move Struggle. 
Added Dark type attacking effectiveness, because of the move Bite. 
And added FAIRY and steel type defending effectiveness, because of some pokemon, like the Jigglipuff line and Magnemite line, being in the base game